+++ lib/brand.ts
@@ -1,0 +1,25 @@
+/**
+ * Application Identity (Brand)
+ *
+ * Also note that the 'Brand' is used in the following places:
+ *  - README.md             all over
+ *  - package.json          app-slug and version
+ *  - public/manifest.json  name, short_name, description, theme_color, background_color
+ */
+export const Brand = {
+  // Name: 'big-AGI',
+  // UpperName: 'BIG-AGI',
+  Title: {
+    Common: 'big-AGI',
+  },
+  Meta: {
+    Description: 'big-AGI is a free, open-source project to build a general artificial intelligence (AGI) that can solve any problem.',
+    Keywords: 'artificial general intelligence, agi, openai, gpt-4, ai personas, code execution, pdf import, voice i/o, ai chat, artificial intelligence',
+    ThemeColor: '#434356',
+  },
+  URIs: {
+    // Slug: 'big-agi',
+    // Deployment: 'https://big-agi.com',
+    OpenRepo: 'https://github.com/enricoros/big-agi',
+  },
+};
