+++ lib/theme.ts
@@ -4,27 +4,6 @@ import { Inter, JetBrains_Mono } from 'next/font/google';
 import { extendTheme } from '@mui/joy';
 
 
-// Identity, Branding & Meta
-
-export const Brand = {
-  // Name: 'big-AGI',
-  // UpperName: 'BIG-AGI',
-  Title: {
-    Common: 'big-AGI',
-  },
-  Meta: {
-    Description: 'big-AGI is a free, open-source project to build a general artificial intelligence (AGI) that can solve any problem.',
-    Keywords: 'artificial general intelligence, agi, openai, gpt-4, ai personas, code execution, pdf import, voice i/o, ai chat, artificial intelligence',
-    ThemeColor: '#434356',
-  },
-  URIs: {
-    // Slug: 'big-agi',
-    // Deployment: 'https://big-agi.com',
-    OpenRepo: 'https://github.com/enricoros/big-agi',
-  },
-};
-
-
 // Theme & Fonts
 
 // For next April Fools' week
