+++ pages/_app.tsx
@@ -6,8 +6,9 @@ import { CacheProvider, EmotionCache } from '@emotion/react';
 import { CssBaseline, CssVarsProvider } from '@mui/joy';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
-import { Brand, createEmotionCache, theme } from '@/lib/theme';
 import '../styles/GithubMarkdown.css';
+import { Brand } from '@/lib/brand';
+import { createEmotionCache, theme } from '@/lib/theme';
 
 
 // Client-side cache, shared for the whole session of the user in the browser.
