+++ pages/_document.tsx
@@ -4,8 +4,9 @@ import { default as Document, DocumentContext, DocumentProps, Head, Html, Main, 
 import createEmotionServer from '@emotion/server/create-instance';
 import { getInitColorSchemeScript } from '@mui/joy/styles';
 
+import { Brand } from '@/lib/brand';
 import { MyAppProps } from './_app';
-import { bodyFontClassName, Brand, createEmotionCache } from '@/lib/theme';
+import { bodyFontClassName, createEmotionCache } from '@/lib/theme';
 
 
 interface MyDocumentProps extends DocumentProps {
